+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 - [feat: (#44) `!` logical NOT operator](https://github.com/amp-lang/amp/issues/44)
 - [feat: (#44) `&&` logical AND operator](https://github.com/amp-lang/amp/issues/44)
 - [feat: (#44) `||` logical OR operator](https://github.com/amp-lang/amp/issues/44)
+- [feat: (#57) `-` negate operator](https://github.com/amp-lang/amp/issues/57)
 
 ### Library
 - feat: implement standard library, I guess